revert short.count to int type

+++ storage/merkle/tree.go
@@ -116,8 +116,8 @@ PutLoop:
 		// node; in that case, we use as much of the shared path as possible
 		case *short:
 			// first, we find out how many bits we have in common
-			commonCount := uint16(0)
-			for i := 0; i < int(n.count); i++ {
+			commonCount := 0
+			for i := 0; i < n.count; i++ {
 				if bitutils.ReadBit(key, i+index) != bitutils.ReadBit(n.path, i) {
 					break
 				}
@@ -128,21 +128,21 @@ PutLoop:
 			// we can simply forward to the child of the short node and continue
 			if commonCount == n.count {
 				cur = &n.child
-				index += int(commonCount)
+				index += commonCount
 				continue PutLoop
 			}
 
 			// if the common count is non-zero, we share some of the path;
 			// first, we insert a common short node for the shared path
 			if commonCount > 0 {
-				commonPath := bitutils.MakeBitVector(int(commonCount))
-				for i := 0; i < int(commonCount); i++ {
+				commonPath := bitutils.MakeBitVector(commonCount)
+				for i := 0; i < commonCount; i++ {
 					bitutils.WriteBit(commonPath, i, bitutils.ReadBit(key, i+index))
 				}
 				commonNode := &short{count: commonCount, path: commonPath}
 				*cur = commonNode
 				cur = &commonNode.child
-				index += int(commonCount)
+				index += commonCount
 			}
 
 			// we then insert a full node that splits the tree after the shared
@@ -151,7 +151,7 @@ PutLoop:
 			var remain *node
 			splitNode := &full{}
 			*cur = splitNode
-			if bitutils.ReadBit(n.path, int(commonCount)) == 1 {
+			if bitutils.ReadBit(n.path, commonCount) == 1 {
 				cur = &splitNode.left
 				remain = &splitNode.right
 			} else {
@@ -166,9 +166,9 @@ PutLoop:
 			// forward to its path; finally, we set the leaf to original leaf
 			remainCount := n.count - commonCount - 1
 			if remainCount > 0 {
-				remainPath := bitutils.MakeBitVector(int(remainCount))
-				for i := 0; i < int(remainCount); i++ {
-					bitutils.WriteBit(remainPath, i, bitutils.ReadBit(n.path, i+int(commonCount)+1))
+				remainPath := bitutils.MakeBitVector(remainCount)
+				for i := 0; i < remainCount; i++ {
+					bitutils.WriteBit(remainPath, i, bitutils.ReadBit(n.path, i+commonCount+1))
 				}
 				remainNode := &short{count: remainCount, path: remainPath}
 				*remain = remainNode
@@ -202,7 +202,7 @@ PutLoop:
 			for i := 0; i < finalCount; i++ {
 				bitutils.WriteBit(finalPath, i, bitutils.ReadBit(key, index+i))
 			}
-			finalNode := &short{count: uint16(finalCount), path: []byte(finalPath)}
+			finalNode := &short{count: finalCount, path: []byte(finalPath)}
 			*cur = finalNode
 			cur = &finalNode.child
 			index += finalCount
@@ -250,7 +250,7 @@ GetLoop:
 		// its paths has all bits in common with the key we are retrieving
 		case *short:
 			// if any part of the path doesn't match, key doesn't exist
-			for i := 0; i < int(n.count); i++ {
+			for i := 0; i < n.count; i++ {
 				if bitutils.ReadBit(key, i+index) != bitutils.ReadBit(n.path, i) {
 					return nil, false
 				}
@@ -258,7 +258,7 @@ GetLoop:
 
 			// forward pointer and index to child
 			cur = &n.child
-			index += int(n.count)
+			index += n.count
 
 			continue GetLoop
 
@@ -330,15 +330,15 @@ ProveLoop:
 		case *short:
 
 			// if any part of the path doesn't match, key doesn't exist
-			for i := 0; i < int(n.count); i++ {
+			for i := 0; i < n.count; i++ {
 				if bitutils.ReadBit(key, i+index) != bitutils.ReadBit(n.path, i) {
 					return nil, false
 				}
 			}
 
 			cur = &n.child
-			index += int(n.count)
-			shortPathLengths = append(shortPathLengths, n.count)
+			index += n.count
+			shortPathLengths = append(shortPathLengths, uint16(n.count))
 			shortNodeVisited = append(shortNodeVisited, true)
 			steps++
 
@@ -431,7 +431,7 @@ DelLoop:
 			last = cur
 
 			// if the path doesn't match at any point, we can't find the node
-			for i := 0; i < int(n.count); i++ {
+			for i := 0; i < n.count; i++ {
 				if bitutils.ReadBit(key, i+index) != bitutils.ReadBit(n.path, i) {
 					return false
 				}
@@ -439,7 +439,7 @@ DelLoop:
 
 			// forward pointer and index to the node child
 			cur = &n.child
-			index += int(n.count)
+			index += n.count
 
 			continue DelLoop
 
@@ -514,12 +514,12 @@ func (t *Tree) Hash() []byte {
 // merge will merge a child short node into a parent short node.
 func merge(p *short, c *short) {
 	totalCount := p.count + c.count
-	totalPath := bitutils.MakeBitVector(int(totalCount))
-	for i := 0; i < int(p.count); i++ {
+	totalPath := bitutils.MakeBitVector(totalCount)
+	for i := 0; i < p.count; i++ {
 		bitutils.WriteBit(totalPath, i, bitutils.ReadBit(p.path, i))
 	}
-	for i := 0; i < int(c.count); i++ {
-		bitutils.WriteBit(totalPath, i+int(p.count), bitutils.ReadBit(c.path, i))
+	for i := 0; i < c.count; i++ {
+		bitutils.WriteBit(totalPath, i+p.count, bitutils.ReadBit(c.path, i))
 	}
 	p.count = totalCount
 	p.path = totalPath
